libdummy-0.1: flatten out directory structure
The example library and programs below libdummy-0.1/ are organized in an
unnecessarily deep directory structure: the library itself and each of the
four programs linking against it are kept in a separate subdirectory each.
This makes working with it cumbersome and hinders parallel compilation.

Flatten out the subdirectory structure below libdummy-0.1/.

Merge the Makefile.am's into one. Note that the objects linked into
the libdummy.so live patch target need extra compilation flags
(-fpatchable-function-entry) and these used to be listed explicitly in a
custom .c.o implicit Make rule. However, this would clash with the one
emitted by Automake for the example programs now. Do not include an
implicit .c.o Make rule in the Makefile.am, but control the resp. objects'
compilations by setting target-specific Make variables for these.

Finally, adapt configure.ac to the new directory structure.

Signed-off-by: Nicolai Stange <nstange@suse.de>