Move the libdummy live patches into libdummy's directory
Move the sources of the live patches to libdummy to the latter's
directory so that everything related to this example is rooted at
a common directory.

Note that in order to not break parallel make, the live patches still need
to be built in their own subdirectory, because the ulp_packer tool doesn't
allow for the specification of an output filename yet.

So, move
  libdummy_livepatch_1-0.1/ to libdummy-0.1/livepatch1/
and
  libdummy_livepatch_2-0.1/ to libdummy-0.1/livepatch2/.

Adapt configure.ac and the Makefile.am's accordingly. Update the README.

Signed-off-by: Nicolai Stange <nstange@suse.de>