Move the libfibo live patch into libfibo's directory
Move the sources of the live patch to libfibo to the latter's directory so
that everything related to this example is rooted at a common directory.

Note that in order to not break parallel make, the live patches still need
to be built in their own subdirectory, because the ulp_packer tool doesn't
allow for the specification of an output filename yet.

So, move libfibo_livepatch-0.1/ to libfibo-0.1/livepatch1/.

Adapt configure.ac and the Makefile.am's accordingly.

Signed-off-by: Nicolai Stange <nstange@suse.de>