Move libdummy + libfibo examples into a subdirectory
Currently, the libdummy-0.1 and libfibo-0.1 examples are located at the
toplevel of libpulp's repository. In order to make clear that these
are examples and to not to clutter the toplevel too much, move them into a
new "examples" directory, i.e. rename
  libdummy-0.1 to examples/dummy
and
  libfibo-0.1 to examples/fibo.

Adapt configure.ac and the Makefile.am's accordingly. Update the README.

Signed-off-by: Nicolai Stange <nstange@suse.de>