Slightly improve usage message for ulp_check and ulp_trigger
The second argument to ulp_check and ulp_trigger is supposed to be the
path to a binary livepatch metadata file, however the usage message is
not absolutely clear about it.  Moreover, the metadata format doesn't
have anything like a magic number, so ulp_check will accept anything,
including DSOs from livepatches, which could make sense according to the
current usage message.