Exit the tools if livepatch metadata was not properly parsed
When reading the livepatch metadata, the expected fields are parsed in a
specific order, then, if anything goes wrong, a warning message is
emitted and an error code is returned.  However, the testing of the
return code on the calling function is wrong and lets the tool keep
going, which is very likely to end in a crash.