Auto generate absolute paths in the metadata description file
The metadata for live patches contains the absolute path to the live
patch DSO and to the library that needs to be patched.  However, since
these paths are absolute, running 'make check' produces binaries that
need to be installed into the paths that are hardcoded in the
description file.

Installing is usually impossible to do as a non-root user, so this patch
converts the description file into a template so that it can be
generated by 'make check' with absolute paths that lay within the build
directory.  With that, the need to install is gone, which allows a
regular user to run the examples.

Now, in order to run the examples from the examples/dummy/ directory,
simply run the usual configure, make, and make check steps, then follow
the instructions below using two different terminals:

From terminal #1, run the following command, which will start printing
timed messages with the pid number of the running process:

  $ LD_PRELOAD=$PWD/lib/.libs/libulp.so examples/dummy/ex1
  libpulp loaded...
  dummyapp: Hello there :))
  [THREAD 9090] example: foo 1
  THIS IS EX1: [THREAD 9090] example: locked behind bars...
  [THREAD 9090] example: foo 3
  THIS IS EX1: [THREAD 9090] example: locked behind bars...
  [THREAD 9090] example: foo 3
  THIS IS EX1: [THREAD 9090] example: locked behind bars...
  [...]

From terminal #2, run the following command, replacing '9090' with the
the process ID reported by the messages on terminal #1:

  $ ./tools/trigger/ulp_trigger 9090 examples/dummy/livepatch1/metadata1.ulp
  ulp: Patching 9090 succesful.

Finally, check that the messages on terminal #1 changed (live patched):

  THIS IS EX1: [THREAD 9090] LIVEPATCHED!
  THIS IS EX1: [THREAD 9090] LIVEPATCHED!
  [...]

These instructions work for all example programs under examples/dummy:
ex1, ex2, ex3, ex4, as well as for all livepatches under that same
directory: livepatch1 and livepatch2.