Move libdummy's livepatch files into its parent directory
A previous patch converted some of the rules in the Makefiles for
libdummy's livepatches from explicit into implicit, which allows their
reuse.  This patch merges those Makefiles.

Apart from the merge, this patch moves the Makefiles into the parent
directory, which allows the use of $(abs_builddir) in the generation of
the livepatch description files, thus removing hardcoded parts on the
paths.  Also, since the livepatches and the test programs (ex1 etc) are
built on the same directory, do not use check-local to request the build
of the livepatch metadata files, instead, add them as dependencies of
the test programs.