Use examples/dummy as a test suite
Since commit ID a49a01817a1c, the test programs under examples/dummy can
be run without the need to be installed.  However, running them had to
be done manually.  This patch adds the infrastructure to run those tests
automatically during 'make check'.

Since testing live patching functionality requires the execution and
interaction of more than one program, the tests must be executed by a
script (wrapper-script), which starts the test program (ex1, ex2, etc.),
tries to patch it (with ulp_trigger), and analyzes the results.

Currently, the wrapper script simply uses the return code of ulp_trigger
to determine if the test case failed or passed, regardless of what the
test programs prints on the output (this could be improved with the use
of other script languages, such as tcl/expect, in the future).