Remove relative paths from dummy's wrapper script
The wrapper script in the dummy example needs to know the path to two
supporting files, libulp.so and ulp_trigger, which are required to run
the test cases.  Before this patch, it relied on the use of ../../ to
find them. After this patch, it uses the full path of the build
directory, which, although not automatically provided by make, can be
manually provided with AM_LOG_FLAGS.