Fix concurrent builds
Since commit ID 55a70c39bd84, when the livepatches for the dummy example
were moved to the same directory, running make check concurrently (-j)
causes the generation of the metadata files to break, with messages
similar to this:

  ../../tools/packer/ulp_packer descr1.txt libdummy.so
  ../../tools/packer/ulp_packer descr2.txt libdummy.so
  mv metadata.ulp metadata1.ulp
  mv metadata.ulp metadata2.ulp
  mv: cannot stat 'metadata.ulp': No such file or directory
  ../../tools/packer/ulp_reverse metadata1.ulp
  make[3]: *** [Makefile:679: metadata2.ulp] Error 1

This patch adds a new argument to ulp_packer and ulp_reverse so that the
output filename can be specified, instead of always writing to the
standard hardcoded filename (metadata.ulp).  The new argument is used in
dummy's Makefile.am, so that the 'mv' command causing the race-condition
can be removed.  It is also used in fibo's Makefile.am, even though it
didn't have concurrency problems.

Also, remove the check for argc != 3 in packer.c, as it provided an
unrelated error message, and would prevent the passing of the third
argument.