Do not ignore the return of read and kill calls
Check the return code of the read call and return on error (returning
here causes the whole call stack to return error codes and, ultimately,
the program (ulp_trigger or ulp_check) to exit.  Likewise, forward the
return code from kill calls, so that the programs can use their return
code and exit on errors.