Fix target function address calculation
The address calculation in get_fentry_from_ulp uses the data in the .ulp
section to obtain the address of the to-be-patched function.  The jump
slots in the .ulp section are composed of three instructions: the first
loads the actual address of the target function into a register (the
address is hardcoded as an immediate offset from the current instruction
point (%rip)); the second pushes the address onto the stack; the last
actually performs the [indirect] jump.

The get_fentry_from_ulp function calculates the same address, however,
it does it without executing the code in the jump-slot.  First, the
desired entry in the .ulp section is obtained with a call to dlsym;
then, just the immediate is read from the first instruction (since the
immediate is in the 4th to 7th bytes of the lea instruction, the number
3 is added to the jump-slot start address); finally, the number 7 is
subtracted from the result, to account for the size of the lea
instruction.

This patch fixes the reading of the immediate from the lea instruction.
The immediate has 4-bytes, however, the memcpy call requests 8 bytes,
which is wrong and doesn't fit in the destination, an int variable.
There's some luck in the fact that the current calculation works.

This patch also removes the second memcpy, because its results are not
used anywhere.

Without this patch, building with -Werror=array-bounds produces the
following warning message:

  lib/ulp.c: In function 'get_fentry_from_ulp':
  lib/ulp.c:150:5: error: 'memcpy' forming offset [5, 8] is out of the
  bounds [0, 4] of object 'offset' with type 'int' [-Werror=array-bounds]
    150 |     memcpy(&offset, func + 3, 8);
        |     ^~~~~~~~~~~~~~~~~~~~~~~~~~~~
  lib/ulp.c:149:9: note: 'offset' declared here
    149 |     int offset = 0;
        |         ^~~~~~
  cc1: some warnings being treated as errors