Treat patch_id as simple sequence of chars and not as a string
The patch_id field (second field in the metadata and member of the
structs ulp_metadata, ulp_applied_patch, and ulp_detour) is a 32-bytes
field, codified as a char array.  Patch IDs, although composed of
characters from the ASCII table are not strings per se, as they never
contain a string termination character.  However, some places in the
code treat it as a string and call strncpy or strncmp, instead of memcpy
or memcmp.

This patch replaces the uses of strncpy and strncmp with memcpy and
memcmp, respectively.  Replacing strncpy with memcpy removes the
following warning:

  lib/ulp.c:885:5: warning: 'strncpy' specified bound 32 equals
  destination size [-Wstringop-truncation]
    885 |     strncpy(detour->patch_id, patch_id, 32);
        |     ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

Also, since all uses of the patch ID just copy or compare bytes, this
patch changes the type specifiers of struct members and function
parameters from char to unsigned char.  In ISO C, the signedness of the
char type is implementation defined and, although libpulp does not care
for signedness of the bytes that compose a patch ID, setting it to
unsigned char conveys the idea that it's not an array of characters nor
a string, but rather an array of byte blobs.  Doing so, removes warning
messages, such as the following:

  lib/ulp.c: In function '__ulp_check_applied_patch':
  lib/ulp.c:96:35: warning: pointer targets in passing argument 1 of
  'ulp_get_applied_patch' differ in signedness [-Wpointer-sign]
     96 |     patch = ulp_get_applied_patch(__ulp_path_buffer);
        |                                   ^~~~~~~~~~~~~~~~~
        |                                   |
        |                                   char *
  In file included from lib/ulp.c:35:
  lib/../include/ulp.h:124:27: note: expected 'unsigned char *' but a
  rgument is of type 'char *'
    124 | struct ulp_applied_patch *ulp_get_applied_patch(unsigned char *id);
        |                           ^~~~~~~~~~~~~~~~~~~~~

Finally, dep_id has the same meaning of patch_id, so it also has its
type specifier changed from char to unsigned char.