Set unused attribute for unused parameter in callback
The dl_iterate_phdr function takes a callback as one of its parameters.
Such callback must have a predefined set of parameters, of which the
second, 'size', is the size of first parameter, 'info'.  According to
the manpage for dl_iterate_phdr, the size parameter should be checked
if struct members other than the pre-GLIBC-2.4 ones, i.e.: dlpi_addr,
dlpi_name, dlpi_phdr, and dlpi_dhnum, are going to be used.  However,
the callback implementation in libpulp only uses the aforementioned
struct members, hence it needs not check the size parameter, which
renders it unused, and produces the following warning message, when
built with -Wunused-parameter:

  lib/ulp.c:733:57: warning: unused parameter ‘size’ [-Wunused-parameter]
    733 | int compare_build_ids(struct dl_phdr_info *info, size_t size, void *data)
        |                                                  ~~~~~~~^~~~

Since it is safe to not check 'size' before using dlpi_addr and friends,
this patch marks the parameter with __attribute__ ((unused)).