Fix file descriptor checks
The main function in dynsym_gate.c and packer.c declare the fd variable
to be used as a regular file descriptor in open and close calls.  The
variable is always initialized to zero before use with open, and should
be checked for greater-than-zero before the call to close.  However, the
pointer to fd is being checked, which is not what was intended.

This patch fix this, thus getting rid of warning messages similar to:

  tools/dynsym_gate/dynsym_gate.c:84:12: warning: ordered comparison of
  pointer with integer zero [-Wextra]
     84 |     if (fd > 0) close(*fd);
        |            ^