Use types that better match the semantics on iterators and indexes
This patch changes the types of variables used as iterators and as array
indexes to types that more closely match the semantics of the operations
they are used in.  Where an exact match is obvious, such as when the
variable is just used in increments and compared against another
variable, the type is changed to be the same of the variable it is
compared against.  When the match is less obvious, for instance when the
variable is used for multiple purposes, only the signedness of the
variable is changed, so that -Wsign-compare does not produce warning
messages similar to the following:

  tools/dynsym_gate/dynsym_gate.c:174:19: warning: comparison of integer
  expressions of different signedness: 'int' and 'size_t' {aka 'long
  unsigned int'} [-Wsign-compare]
    174 |     for (i = 0; i < nr; i++) {
        |                   ^

(in the main function in dynsym_gate.c, 'i' is used both in operations
with multiple other variables, which do not match exactly in type,
unsigned int and size_t, but both have unsigned semantics)

Once all compilation warnings are gone, -Wall, -Wextra, and -Werror will
be added to the default build options to avoid regressions.