Fix maybe-uninitialized warnings
Building with -Wmaybe-uninitialized produces several warnings similar to
the following:

  lib/ulp.c:634:22: warning: 'prev_unit' may be used uninitialized in
  this function [-Wmaybe-uninitialized]
    634 |      prev_unit->next = a_unit;
        |      ~~~~~~~~~~~~~~~~^~~~~~~~

Most of them are caused by linked list code that works perfectly well,
as the initialization is based on another variable that is definitely
set, but the compiler is unable to detect the indirection.  For these
cases, this patch simply initializes the linked list pointers to NULL,
so even if the code was wrong, a segmentation fault (null dereference)
would happen, instead of a write to an unknown location.  Other cases
are discussed below:

 - The for loop over 'j' (in main in dynsym_gate.c) was using 'i' to
   access the data in the array, a common typo, which, in this case,
   never exposed a bug.

 - The 'dyn_addr' variable (in dig_main_link_map in introspection.c) is
   used as an accumulator, so it must be zero-initialized.

 - The 'entry' and '_start' variables (in dig_load_bias in
   introspection.c), now renamed to 'addrof_entry' and 'addrof_start',
   should be found with the introspection into the to-be-patched
   process, if they are not found, something went wrong, so this patch
   initializes these variables to zero and verifies that they hold
   something else afterwards.

 - In packer.c, 'goto main_error' statements can take control-flow to
   the de-initialization code, where fd gets closed, however, the first
   of the gotos does it before fd is ever initialized.