Signal libulp initialization to external programs
Although not a common case, when a live patch is triggered too soon,
more specifically, when it is triggered during the loading of the target
process, it will likely fail to apply.  This situation has not been
observed to happen with a one second delay between the target process
initialization and the triggering of the patch, i.e.: when SLEEP_AMOUNT
(examples/dummy/wrapper-script) is set to one.  Anyhow, this patch
addresses the aforementioned situation.

First, it adds a global variable, __ulp_state.load_state, to libulp,
which is initially set to zero, and will only be set to one in the
library constructor (begin).  Since constructors are only invoked after
all shared objects dependencies have already been loaded by ld.so, this
signals, to ulp_trigger and ulp_check, the time when it is safe to
actually patch the process.

Second, this patch changes the test wrapper in examples/dummy so that it
now retries to apply the patch if the live patching failed due to
initialization reasons (i.e. with exit codes 4 or EAGAIN).  EAGAIN means
that libulp is not fully initialized (the constructor has not yet been
run), whereas 4 means that the call to initialize_data_structures failed
for some other reason, yet, retrying later might succeed.  In both
cases, retrying again later is fine.

Without the check for full initialization, trying to actually apply the
patch might leave the target program in an inconsistent state.  The
following is a non-comprehensive list of symptoms observed before this
patch, and fixed by it (symptoms #5, #6 and #7 show situations when the
target program was put into an inconsistent state):

  Symptom #1:
  ulp: error: unable to find entry address for the executable
  Live patch #1 failed to apply.

  Symptom #2:
  ulp: error: unable to find address for _start
  Live patch #1 failed to apply.

  Symptom #3:
  ulp: ulp loop address not found.
  Live patch #1 failed to apply.

  Symptom #4:
  ulp: read_memory error.
  ulp: error reading link_map address.
  Live patch #1 failed to apply.

  Symptom #5:
  ulp: Unable to read byte (write).
  ulp: set_path_buffer error 2.
  ulp: Apply patch to 12608 failed.
  ulp: PTRACE_ATTACH error.
  ulp: Restore threads error (can't attach).
  Live patch #1 failed to apply.

  Symptom #6:
  ulp: Patching 20285 succesful.
  ulp: read_memory error.
  ulp: error reading _DYNAMIC array.
  wrapper-script: line 48: 20285 Segmentation fault
    (core dumped) LD_PRELOAD=$builddir/lib/.libs/libulp.so $testcase
  Live patch #2 failed to apply.
  wrapper-script: line 60: kill: (20285) - No such process

  Symptom #7:
  ulp: Patching 2073 succesful.
  ./ex2: Relink `/lib64/libdl.so.2' with `/lib64/libc.so.6' for IFUNC symbol `strcpy'
  ulp: read_memory error.
  ulp: error reading _DYNAMIC array.
  wrapper-script: line 48:  2073 Segmentation fault
    (core dumped) LD_PRELOAD=$builddir/lib/.libs/libulp.so $testcase
  Live patch #2 failed to apply.
  wrapper-script: line 60: kill: (2073) - No such process