Fix uses of the global universe counters
The code consistency model of libpulp enables live patches to be
applied, regardless of the existence of multiple threads and multiple
live-patchable (target) libraries in a process, and it works as follows:
each target library has its own global counter, and each thread has its
own counters, one for each of the target libraries; whenever a live
patch is applied, the per-library global counter of the library being
live patched gets incremented; however, this does not immediately
trigger the use of the new functions by the threads of the process,
because each thread must, first, increment its thread counters to match
the global counters, and that only happens on library calls (i.e. when
the thread enters the library).  This guarantees that the whole
execution of a library call, including internal calls to other functions
within the library, happens within function versions that are consistent
with each other, i.e. belonging to the same live patch snapshot.

When a live patch is applied, ulp_trigger (the tool) accesses code in
libulp.so, which must have been LD_PRELOAD'ed during the initialization
of the process.  However, the code in libulp.so currently makes by-name
accesses to the aforementioned global counter (__ulp_global_universe).
It also has its own global counter, which gets incremented, instead of
the target library's global counter mentioned in the first paragraph.

Such mechanism, which would have a single, per-process global counter,
could theoretically work, since the global universe is always
incremented by an external tool and such increments could be externally
coordinated to never occur simultaneously, thus not causing a race
condition.  However, this patch removes the global counter from
libulp.so, as well as the by-name references to it, replacing them with
calls to dlsym.

Having a global counter per library has an additional advantage:
__ulp_entry, in trm.S, is executed every time the library is entered,
and it needs to read the global counter (__ulp_global_universe) to,
potentially, update the thread counter (__ulp_thread_universe); with a
per library global counter, the read location is known at link time (it
is even in the same compilation unit), which makes the access very
simple, and hopefully fast.  If, instead, we were to have a single
global counter per process (in libulp.so), some other access mechanism,
probably involving calls to dlopen and dlsym, would be needed in
__ulp_entry, potentially and very likely leading to performance hits.