Test multiple target libraries
The test driver for the test programs in the examples directory only
verifies the return code of the ulp_trigger command execution.  This
provides some confidence that a live patch applied correctly, but not
fully, because it lacks a validation of the new behavior expected of the
target library.

This patch adds a new test case to make check, one that actually
validates the new behavior of a live patched library, hence the use of a
new subdirectory, tests.  Apart from that, the new test case applies
live patches to multiple (currently two) target libraries.