Fix metadata parsing
The metadata file stores strings without the terminating null byte.
Each string is preceded by a 4-byte length field so that it can be
correctly parsed.  However, some parts of libpulp do not account for the
additional char required to terminate the string, which might lead to
parsing problems.

I have observed such problem happenning on rare occasions, as a
side-effect of having some specific file path length, for instance:

  $ cd /tmp/0123456789012345678901234567890123456789
  $ <libpulp path>/configure && make && make check
  $ cat examples/dummy/ex1.log
  (snip)
  ulp: to be patched object (/tmp/0123456789012345678901234567890123456789/examples/dummy/libdummy.so1) not loaded.

  (notice that the string goes beyond the actual filename)

This patch adds the extra character to the variables that read strings
from the metadata.  It does not explicitly set the last char to the
terminating null byte, because calloc already initializes the memory to
all zeroes.