Convert __ulp_global_universe to pointer access
The commit

commit 2dd03cec1cbbc23d2866bd918571045d3858810e
Author: Gabriel F. T. Gomes <gabriel@inconstante.net.br>
Date:   Mon Mar 16 13:24:28 2020 -0300

    Fix uses of the global universe counters

was based on the assumption that each target library should have its
own, per-library, global counter (__ulp_global_universe), because it
would make it easier for such libraries to access them.  However, the
assumption was wrong, and the initial design of libpulp had a single,
per-process, global counter, which would be accessed by each library via
the GOT, so the commit is reverted.

Despite the fact that trm.S declares a __ulp_global_universe object and
that trm.o gets linked with each target library, this per-library object
is not normally accessed, because libulp.so is LD_PRELOAD'ed, and its
global counter takes precedence over those in the target libraries.

However, when a target library is linked with -Bsymbolic, the object
within the library takes precedence, as described in ld's manual:

  -Bsymbolic
  When creating a shared library, bind references to global symbols to
  the definition within the shared library, if any. Normally, it is
  possible for a program linked against a shared library to override the
  definition within the shared library.

This breaks the libpulp model and has been observed to be used at least
in openssl libraries: libssl and libcrypto.

This patch refactors the per-process global counter so that it exists
only in libulp, but has references to it in all target libraries.  Such
by-reference access needs to be set up during the process startup, so
libulp's constructor initializes the references.  If libulp is not
LD_PRELOAD'ed, the references are not initialized and the universe
synchronization code is skipped.

This mechanism restores libpulp's per-process global counter and works
with -Bsymbolic linking.