Makefile and configure cleanup
Makefile rules to build the examples and tests caused some Makefile code
duplication.  This patch moves the common code to Makefile.common, as
well as it converts the dummy and fibo examples to more thoroughly using
libtool, similar to what the libraries and test case in the tests
directory already do.

This patch also moves automake's --foreign option from the bootstrap.sh
script into configure.ac, which enables the use of 'autoreconf -i'.
Finally, it drops the .sh extension from bootstrap.sh, which is more
inline with automake's manual (the actual recommendation is to drop
bootstrap or autogen.sh scripts altogether, in favor of autoreconf).