Allow use of -Bsymbolic in target library linking
The commit

commit 8f628ab7855745fc585fd29727b022bb938c0790
Author: Gabriel F. T. Gomes <gagomes@suse.de>
Date:   Tue Mar 24 21:29:03 2020 -0300

    Convert __ulp_global_universe to pointer access

has been reverted in favor of the simpler, and a lot less likely to
cause undesired side-effects, approach implemented by this patch.

This patch removes, from trm.S, the definition of __ulp_global_universe,
hence no object is created within the target library.  This guarantees
that, during linking, the reference to __ulp_global_universe will never
be resolved to an address within the target library, even when
-Bsymbolic is passed to the linker.  Instead, the reference is left
unresolved as a GOT-based reference, which may or may not be resolved by
the loader, depending on whether libulp.so has been LD_PRELOAD'ed or
not.  If it has, symbol relocation finds the address of the global
counter in libulp.so and fills the GOT slot accordingly; otherwise, the
GOT slot is left uninitialized, which works fine nonetheless, because,
in this case, user-space livepatching will not happen anyway.