Check linker support for the .ulp section
Libraries that are to be live-patchable must be linked against trm.S so
that they have a __ulp_entry function, as well as trampolines that
exported functions go through on library entrance.  Such trampolines are
placed into the .ulp section, which gets created by the linker when a
.ulp.track section exists in the objects being linked.  Since the
creation of the .ulp section requires specific linker support, check for
its availability during the configure step and fail early with a
friendly warning, if it is not available.