Test multiple parameters in live patched function
None of the functions that get live patched in the examples and tests
have parameters.  This patch adds a new test case, which tries to live
patch a function with four integer parameters and fails, because the
parameters are not properly saved and restored by the universe
management functions.  The test is marked as expected to fail.