Fix saving and restoring of integer and pointer parameters
The ABI for AMD64 states that integer and pointer parameters are to be
passed in the following registers: %rdi, %rsi, %rdx, %rcx, %r8, and %r9.
However, when a function that has been livepatched is called, the live
patch detour only saves %rdi before jumping into __ulp_prologue, which
then calls __ulp_manage_universes, a function written in C.

Except for the recently added 'parameters' test, all the examples and
tests in libpulp apply live patches to functions that have no
parameters, so they get away with the problem described above.

This patch fixes parameter saving and restoring for integer and pointer
parameters.  Other classes of parameters are still broken and probably
need fixing.