Move __ulp_loop from target libraries to libulp
In order for live patching of a process to work, libulp provides some
functions that enable an external process to jack-in.  Among these
functions, __ulp_loop is the actual entry point, and it gets executed by
all threads in the to-be-patched process, after it has been hijacked.

However, __ulp_loop currently resides on trm.S, which means that every
live-patchable library has a copy of it, thus the to-be-patched process
has as many __ulp_loop functions as the number of live-patchable
libraries it is linked against.  This is not required and somewhat
misleading, as it might give the impression that a live patch to some
library must use __ulp_loop from that same library.

This patch moves __ulp_loop to libulp.so, which removes the duplication.