Userspace live patching tool
This patch adds an initial sketch of what a ulp tool would look like,
particularly, it adds the functionaly that enables introspection tools
to read the global and thread-local universe counters. Being able to
read the thread-local universe counter is important to determine if an
already installed live patch is actually in use. If any thread-local
universe counter has a value that is less than the global universe (a
situation that may happen if a thread never leaves a library, but loops
or sleeps within it) the live patch, although installed, is not in use.