Move process stop and restart into the hijacking functions
Before all the threads in a process are put into an infinite loop
(__ulp_loop), they are all stopped with a SIGSTOP signal (sending
SIGSTOP to a process stops all of its threads). After the operation,
they are all resumed by sending a SIGCONT signal to the process.  The
same thing applies when the threads are removed from the infinite loop.

This patch moves the sending of the SIGSTOP and SIGCONT signals into
these thread hijacking and restoring functions. It also fixes the return
code of the main function in check.c, which is used in dispatcher.lua to
determine if a live patch has been applied to a process.