Test blocked threads
Libpulp's model of consistency is based on the fact that each thread
only observes the changes caused by a live patch on library entrance,
i.e. even after a live patch has been successfully applied to a process,
the new function will only be executed by a thread after the thread
leaves the library and enter again via a new library call. In other
words, a thread that is looping within the target library will never be
actually live-patched.

This patch adds a test case that demonstrate this behavior. This test
case can also be used with the ulp tool (tools/trigger/ulp.c) to see how
the introspection tool reports such blocked threads, for instance:

  $ tools/trigger/ulp
  PID: 23840
    Global universe: 1
    Live patches:
      /<snip>/tests/.libs/libblocked_livepatch1.so
    Local universes:
      in /<snip>/tests/.libs/libblocked.so.0:
        thread #023858: OK (queued - not currently in the library)
        thread #023859: NOK (blocked - thread-local universe=0)
        thread #023840: OK (queued - not currently in the library)