Add internal allocator
During livepatch installation, libulp allocates lots of small chunks of
memory. Some of these chunks get added to internal structures of live
patching control, such as the list of detours and their lists of applied
patch units; these chunks never get freed. Whereas other chunks have a
short life and are only used during metadata parsing, then freed.

Also, since libulp functions are executed from the context of a signal
handler, they are suppose to only make calls to assynchronous signal
safe (AS-Safe) functions. However, malloc and calloc are AS-Unsafe, so
they should not be used, yet, they are currently used in the allocations
mentioned in the first paragraph, which might lead to deadlocks, as
demonstrated by the deadlock test case (tests/deadlock).

This patch adds an internal memory allocator to libulp that can be used
instead of malloc/calloc. Since the requested chunks of memory are
either permanent (never freed) or short-lived (freed within the signal
handler), as described in the first paragraph, this allocator has two
pools (arenas) that take this allocation strategy into account.

The default arena is used for permanent data and is supposed to never
ever be freed, whereas the scratch arena may be freed, however, it only
allows the freeing of the whole arena, not of individual chunks. This
mechanism allows this allocator to be a lot simpler than malloc.

To use the new memory allocator, libulp must choose the appropriate
alloc function, depending on the use semantics. For the permanent chunks
of memory, it should call ulp_default_alloc, whereas for the temporary
chunks used in metadata parsing, it should call ulp_scratch_alloc.
Finally, before leaving the signal handler context, libulp should call
ulp_scratch_clear to clear the whole scratch arena.