Use internal allocator in libulp
Now that libulp has an internal allocator, replace all calls to calloc
and free with calls to: ulp_default_alloc, when allocating data that
goes into the detour structures; ulp_default_alloc, for metadata
parsing; and ulp_scratch_clear, instead of freeing individual blocks.
Notice that the calls to malloc and calloc in the tools, for instance,
ulp_trigger, remain unchanged. Only calls that happen from the context
of a signal handler need replacing.

This fixes the deadlock caused by calls to calloc during live patch
installation. Without this patch, the call stack from a hang looks
similar to the following (notice that the deadlock is caused by a call
to __libc_calloc from load_metadata):

  >>> bt
  #0  __lll_lock_wait_private (futex=futex@entry=0x7f62681059e0 <main_arena>) at ./lowlevellock.c:35
  #1  0x00007f6267fd1e87 in __libc_calloc (n=<optimized out>, elem_size=<optimized out>) at malloc.c:3399
  #2  0x00007f626815397e in load_metadata () at lib/ulp.c:224
  #3  0x00007f62681543ce in load_patch () at lib/ulp.c:430
  #4  0x00007f6268153517 in __ulp_apply_patch () at lib/ulp.c:75
  #5  0x00007f626815602c in __ulp_trigger () at lib/ulp_interface.S:35
  #6  0x000000001ff02c00 in ?? ()
  #7  0x00007ffdd072a260 in ?? ()
  #8  0x00000000005f83e0 in ?? ()
  #9  0x0000000000000000 in ?? ()

After this patch, the deadlock is still present, but due to calls to
other AS-Unsafe functions: fopen, fread, fclose, dlopen, and dlsym.
These will be fixed in future patches.