Do not use use fopen, fread, and fclose in libulp.so
The file stream functions: fopen, fread, and fclose are AS-Unsafe, so
they are not safe to call from the signal handler that libpulp uses to
install live patches. This patch removes their use from this scenario,
which fixes the deadlock caused by them. Without this patch, the call
stack from a hang looks similar to the following (notice that the
deadlock is caused by a call to __fopen_internal from parse_metadata):

  >>> bt
  #0  __lll_lock_wait_private (futex=futex@entry=0x7f1e6f4449e0 <main_arena>) at ./lowlevellock.c:35
  #1  0x00007f1e6f30ffeb in __GI___libc_malloc (bytes=bytes@entry=472) at malloc.c:3064
  #2  0x00007f1e6f2faa7b in __fopen_internal (filename=0x7f1e6f4982a0 <__ulp_path_buffer> "libblocked_livepatch1.ulp", mode=0x7f1e6f49519f "rb", is32=1) at iofopen.c:58
  #3  0x00007f1e6f4929bb in parse_metadata (ulp=0x7f1e6f4c6000) at lib/ulp.c:239
  #4  0x00007f1e6f492955 in load_metadata () at lib/ulp.c:221
  #5  0x00007f1e6f493310 in load_patch () at lib/ulp.c:420
  #6  0x00007f1e6f492537 in __ulp_apply_patch () at lib/ulp.c:77
  #7  0x00007f1e6f494f54 in __ulp_trigger () at lib/ulp_interface.S:35
  #8  0x0000000000000000 in ?? ()

After this patch, the deadlock is still present, but due to calls to
other AS-Unsafe functions: dlopen, and dlsym.