Remove hack to cause deadlock from the deadlock test
The deadlock test case (tests/deadlock.c) has a signal handler that
intentionally calls calloc. The test wrapper (tests/deadlock.py)
triggers this signal handler, in a loop, to demonstrate the deadlock
even if live patch installation luckily succeeded (the deadlock is not
always reached during live patch installation).

This patch replaces this demonstration hack with an outer loop of live
patch installations, which increases the chance of hitting a deadlock
during live patch installation. The loop iterates a fixed (hardcoded)
number of times, which is large enough to always reproduce the problem
(at least according to the tests I made).