Add a semaphore to the internal allocator
The internal allocator is only used in code paths that start at
__ulp_trigger and __ulp_check_patched, i.e. only triggered by external
tools and from a signal handler context. These tools always hijack all
the threads in the target process before actually accessing the
aforementioned code paths, and thread hijacking should be started by a
single [external] process at a time.

Nevertheless, the internal malloc structures could, theoretically, be
accessed by multiple threads, which would lead to a race-condition in
the access to the arenas. This patch adds a semaphore to each of the
alloc arenas, which prevents this race-condition.