Work around the deadlock
Add a new interface to libpulp that allows testing if calloc and dlopen
are safe to call from a signal handler, even though they are AS-Unsafe.
Then use it in the trigger tool before actually applying a live patch.
This relies on two functions from glibc, which are not upstream:
__libc_ulp_checks and __libdl_ulp_checks.

Since these functions are not upstream, they will be missing from most
installations, hence the new configure check for their presence.  When
they are not available, libpulp simply ignores the problem and applies
the live patch, which will very likely lead to the usual deadlock, which
now causes a make check failure (deadlock.py is no longer under
XFAIL_TESTS). When they are available, the trigger tool makes calls to
them and only really applies the live patch when it is safe to do so,
i.e. only when all of the related locks are free; when this happens,
live patch application succeeds and the process never goes into the
deadlock.  The glibc patch to add these functions is available at
patches/glibc/ulp-checks.patch.

On an system with unpatched glibc, the test case program (deadlock.c)
exposes deadlocks related to both malloc and dlopen implementations.
When a deadlock situation occurs, attaching to the target process with
gdb and printing the call stacks reveals which lock caused the problem,
for instance:

  Deadlock on malloc implementation (see call to calloc in frame #1):

    (gdb) bt
    #0  0x00007f2622e54e37 in __lll_lock_wait_private () from /lib64/libc.so.6
*   #1  0x00007f2622e5b2a3 in calloc () from /lib64/libc.so.6
    #2  0x00007f2622fbeb18 in _dlerror_run () from /lib64/libdl.so.2
    #3  0x00007f2622fbe2e4 in dlopen@@GLIBC_2.2.5 () from /lib64/libdl.so.2
    #4  0x00007f2622fcc780 in load_so (obj=0x7f2623000050 "/root/build/x86_64/libpulp/tests/.libs/libblocked_livepatch1.so")
        at /root/src/libpulp/lib/ulp.c:463
    #5  0x00007f2622fcb99a in load_so_handlers (ulp=0x7f2623000000) at /root/src/libpulp/lib/ulp.c:264
    #6  0x00007f2622fcc751 in parse_metadata (ulp=0x7f2623000000) at /root/src/libpulp/lib/ulp.c:454
    #7  0x00007f2622fcbaff in load_metadata () at /root/src/libpulp/lib/ulp.c:304
    #8  0x00007f2622fcc7dc in load_patch () at /root/src/libpulp/lib/ulp.c:478
    #9  0x00007f2622fcb628 in __ulp_apply_patch () at /root/src/libpulp/lib/ulp.c:99
    #10 0x00007f2622fce542 in __ulp_trigger () at /root/src/libpulp/lib/ulp_interface.S:44
    #11 0x00007f2622f8f7e0 in ?? () from /lib64/libc.so.6
    #12 0x00007ffe1ce89930 in ?? ()
    #13 0x00007f2622f8f7e0 in ?? () from /lib64/libc.so.6
    #14 0x0000000000000000 in ?? ()

  Deadlock on dlopen implementation (see call to _dl_open in frame #2)

    (gdb) bt
    #0  0x00007fbc87fb3f5c in __lll_lock_wait () from /lib64/libpthread.so.0
    #1  0x00007fbc87fac8f1 in pthread_mutex_lock () from /lib64/libpthread.so.0
*   #2  0x00007fbc87ff5b06 in _dl_open () from /lib64/ld-linux-x86-64.so.2
    #3  0x00007fbc87fc9258 in dlopen_doit () from /lib64/libdl.so.2
    #4  0x00007fbc87f14960 in _dl_catch_exception () from /lib64/libc.so.6
    #5  0x00007fbc87f14a1f in _dl_catch_error () from /lib64/libc.so.6
    #6  0x00007fbc87fc9a65 in _dlerror_run () from /lib64/libdl.so.2
    #7  0x00007fbc87fc92e4 in dlopen@@GLIBC_2.2.5 () from /lib64/libdl.so.2
    #8  0x00007fbc87fd7780 in load_so (obj=0x7fbc8800b050 "/root/build/x86_64/libpulp/tests/.libs/libblocked_livepatch1.so")
        at /root/src/libpulp/lib/ulp.c:463
    #9  0x00007fbc87fd699a in load_so_handlers (ulp=0x7fbc8800b000) at /root/src/libpulp/lib/ulp.c:264
    #10 0x00007fbc87fd7751 in parse_metadata (ulp=0x7fbc8800b000) at /root/src/libpulp/lib/ulp.c:454
    #11 0x00007fbc87fd6aff in load_metadata () at /root/src/libpulp/lib/ulp.c:304
    #12 0x00007fbc87fd77dc in load_patch () at /root/src/libpulp/lib/ulp.c:478
    #13 0x00007fbc87fd6628 in __ulp_apply_patch () at /root/src/libpulp/lib/ulp.c:99
    #14 0x00007fbc87fd9542 in __ulp_trigger () at /root/src/libpulp/lib/ulp_interface.S:44
    #15 0x00007fbc87f9a7e0 in ?? () from /lib64/libc.so.6
    #16 0x00007fff2782e340 in ?? ()
    #17 0x00007fbc87f9a7e0 in ?? () from /lib64/libc.so.6
    #18 0x0000000000000000 in ?? ()

On a system with patched glibc, these deadlocks never occur.

Note: deadlock.c now has 50 threads, which is a frail magic number that
has been observed to cause a deadlock when __libc_ulp_checks or
__libdl_ulp_checks are *not* available.